refactor(bft): widen consensus round from u16 to u32
The BFT `Round` type alias and the coupled `bft_rounds_per_session`
config field were `u16`, capping a session at 65535 rounds. Widen both
to `u32` so session length isn't bounded by the round counter.

+++ picomint-bft/src/unit.rs
@@ -8,7 +8,7 @@ use picomint_encoding::{Decodable, Encodable};
 /// Round number within a session. Round 0 is the first row of the DAG;
 /// its units carry empty parent sets and are otherwise created and
 /// disseminated like every other unit.
-pub type Round = u16;
+pub type Round = u32;
 
 /// Type alias for the trait bound every consumer of `D` ends up
 /// repeating. Anything that round-trips on the wire, can be moved
